+++ christmas_app/static/index.js
@@ -1,3 +1,6 @@
+// Helper functions
+
+// Welcome popup
 function hit() {
   Swal.fire({
     title: "Surprise Present is ready for you!",
@@ -10,6 +13,7 @@ function hit() {
   });
 }
 
+// Pre-defined account checker
 async function theChecker(loc = "") {
   if (loc == "login-integrated") {
     var htmlCode =
@@ -30,6 +34,32 @@ async function theChecker(loc = "") {
   }
   Swal.fire({
     title: "Submit your First Name",
+    html: htmlCode,
+    showCancelButton: true,
+    showConfirmButton: false,
+    showLoaderOnConfirm: true,
+
+    allowOutsideClick: () => !Swal.isLoading(),
+  });
+}
+
+// Account password changer
+async function theChanger(alias = "") {
+  if (alias != "") {
+    var htmlCode =
+      '<form class="form" action="/' +
+      alias +
+      '/account-management/change-password/" method="POST"> ' +
+      "<h3 style='color:red;' >[]</h3>" +
+      '<input class="col-8 my-2" type="password" placeholder="Current Password" aria-label="Search" autocapitalize="characters" name="account_" id="account_"/><br>' +
+      '<input type="submit" value="Confirm Changes" class="btn-lg" />' +
+      "</form>";
+  } else {
+    var htmlCode =
+      "You can't change account password when you didn't login to the system";
+  }
+  Swal.fire({
+    title: "Account Password Changer",
     html: htmlCode,
     showCancelButton: true,
     showConfirmButton: false,