+++ changelog.md
@@ -29,3 +29,5 @@ Config the System version in **init**.py
 - Initial Development#6.1: minor overall improvements on December 11, 2022 -> **0.3221**
 
 - Initial Development#6.2: game implementation started and overall improvements on December 11, 2022 -> **0.4**
+
+- Initial Development#7: session timeout implemented and ss started, login system addon feature (redirect option), and overall improvements on December 12, 2022 -> **0.42**